+++ patches/src/main/kotlin/app/revanced/patches/youtube/video/playback/Fingerprints.kt
@@ -1,5 +1,6 @@
 package app.revanced.patches.youtube.video.playback
 
+import app.revanced.patcher.extensions.InstructionExtensions.instructionsOrNull
 import app.revanced.util.containsLiteralInstruction
 import app.revanced.util.fingerprint.legacyFingerprint
 import app.revanced.util.getReference
@@ -7,6 +8,7 @@ import app.revanced.util.indexOfFirstInstruction
 import app.revanced.util.or
 import com.android.tools.smali.dexlib2.AccessFlags
 import com.android.tools.smali.dexlib2.Opcode
+import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
 import com.android.tools.smali.dexlib2.iface.reference.FieldReference
 
 internal val av1CodecFingerprint = legacyFingerprint(
@@ -54,6 +56,22 @@ internal val playbackSpeedChangedFromRecyclerViewFingerprint = legacyFingerprint
     }
 )
 
+// Fingerprint for the METHOD that returns PlayerConfigModel
+private const val PCM_GETTER_FIELD_TYPE = "Lcom/google/android/libraries/youtube/innertube/model/media/PlayerConfigModel;"
+val pcmGetterMethodFingerprint = legacyFingerprint(
+    name = "pcmGetterMethodFingerprint",
+    returnType = PCM_GETTER_FIELD_TYPE,
+    parameters = listOf(),
+    opcodes = listOf(Opcode.IGET_OBJECT, Opcode.RETURN_OBJECT),
+    customFingerprint = custom@{ method, _ ->
+        val instructions = method.instructionsOrNull
+        if (instructions == null || instructions.count() != 2) return@custom false
+
+        ((method.instructionsOrNull?.firstOrNull() as? ReferenceInstruction)?.reference
+                as? FieldReference)?.type == PCM_GETTER_FIELD_TYPE
+    }
+)
+
 internal val loadVideoParamsFingerprint = legacyFingerprint(
     name = "loadVideoParamsFingerprint",
     returnType = "V",
@@ -73,6 +91,18 @@ internal val loadVideoParamsParentFingerprint = legacyFingerprint(
     parameters = listOf("J"),
     strings = listOf("LoadVideoParams.playerListener = null")
 )
+
+// internal val playbackSpeedSetterFingerprint = legacyFingerprint(
+//     name = "playbackSpeedSetterFingerprint",
+//     accessFlags = AccessFlags.PUBLIC or AccessFlags.FINAL,
+//     returnType = "V",
+//     parameters = listOf("F", "Z"),
+//     opcodes = listOf(
+//         Opcode.IGET_OBJECT,
+//         Opcode.IF_NEZ,
+//         Opcode.GOTO_16,
+//     )
+// )
 
 internal val qualityChangedFromRecyclerViewFingerprint = legacyFingerprint(
     name = "qualityChangedFromRecyclerViewFingerprint",